Clearing mongo database
I fixed my code by adding in the logout button. We added code to take in the json array of the data in alphabetical order to put in the mongo database. We also added another json object that is a function that pulls up friends. We also added a json function that deletes and clears the database.