Creating an IIFE
I found an error in my code in the water fall of an incorrectly spelt string in the json return. I also corrected another error in my index on the id friends-column.We added a file called main.js the file is used to create an iife that gets the user id of a certain person and creates a click event. The click event is triggered when a person has been selected and will eventually allow for notes to be added to the specifically selected user.