Creating notes and changing the url
I fixed some spelling errors in my comments. In the main.js we added a getNotes function which has it's scope stuck in the iffe. So no conflicts will arise. we created document fragment variables to create a mini doc holding the user gui. We created a new object that has an event handler that looks for a click and changes the url to hold the twitter id. Then we made a create notes function that should return the notes.