Updating notes
In the index.ejs file I beautified it. In the storage,js file we added another json function which has the purpose of updating the function. the update requires a parameter that filters for the record we are looking for. it also puts a json data into a variable. We also added event listeners for a blur event and a keydown event. Then we made a put note function that gets a request that is a post and sends it back creating a unique url. in the index.js we created a new route that passes in all of our information into the storage.js.